refactor(lib): split lib into a few files
