+++ src/pathogena/constants.py
@@ -6,17 +6,16 @@ DEFAULT_HOST = "portal.eit-pathogena.com"
 DEFAULT_APP_HOST = "app.eit-pathogena.com"
 DEFAULT_UPLOAD_HOST = "api.upload.eit-pathogena.com"
 DEFAULT_PROTOCOL = "https"
-DEFAULT_METADATA = {
-    "country": None,
-    "district": "",
-    "subdivision": "",
-    "instrument_platform": "illumina",
-    "pipeline": "mycobacteria",
-    "ont_read_suffix": ".fastq.gz",
-    "illumina_read1_suffix": "_1.fastq.gz",
-    "illumina_read2_suffix": "_2.fastq.gz",
-    "max_batch_size": 50,
-}
+DEFAULT_COUNTRY: None = None
+DEFAULT_DISTRICT = ""
+DEFAULT_SUBDIVISION = ""
+DEFAULT_INSTRUMENTPLATFORM = "illumina"
+DEFAULT_PIPELINE = "mycobacteria"
+DEFAULT_ONT_READ_SUFFIX = ".fastq.gz"
+DEFAULT_ILLUMINA_READ1_SUFFIX = "_1.fastq.gz"
+DEFAULT_ILLUMINA_READ2_SUFFIX = "_2.fastq.gz"
+DEFAULT_MAX_BATCH_SIZE = 50
+
 HOSTILE_INDEX_NAME = "human-t2t-hla-argos985-mycob140"
 
 DEFAULT_CHUNK_SIZE = int(
